Add offset to metadata for msgs from imfile
The offset of the beginning of the log line is now added as metadata, if
additional metadata is requested.

There is one odd case, where a partial line is read from a file an EOF
is hit before the proper end of the msg is encountered (for any mode or
multiline msgs).  In that case, the offset of the remaining record in
the subsequent file is used for the entire msg.

We also take this opportunity to consistently use off64_t instead of a
combination of int64, off_t, and off64_t (allowing consistent behavior
in 32 bit environments).

+++ runtime/obj.c
@@ -609,7 +609,7 @@ rsRetVal objDeserializeProperty(var_t *pProp, strm_t *pStrm)
 	number_t iLen;
 	uchar c;
 	int step = 0; /* which step was successful? */
-	int64 offs;
+	off64_t offs;
 
 	assert(pProp != NULL);
 
@@ -666,8 +666,8 @@ rsRetVal objDeserializeProperty(var_t *pProp, strm_t *pStrm)
 finalize_it:
 	if(Debug && iRet != RS_RET_OK && iRet != RS_RET_NO_PROPLINE) {
 		strm.GetCurrOffset(pStrm, &offs);
-		dbgprintf("error %d deserializing property name, offset %lld, step %d\n",
-			  iRet, offs, step);
+		dbgprintf("error %d deserializing property name, offset %llu, step %d\n",
+			  iRet, (long long unsigned)offs, step);
 		if(step >= 1) {
 			dbgprintf("error property name: '%s'\n", rsCStrGetSzStrNoNULL(pProp->pcsName));
 		}
